THORN-2350: upgrade to fraction plugin 94
Motivation
----------
Fraction plugin version 94 fixes THORN-2349.

Modifications
-------------
Upgrades to fraction plugin version 94.

In addition to that, this also fixes the template of an artificial
Maven project used for generating an offline Maven repo. It used
to hardcode a dependency on `io.thorntail:fraction-metadata`, but
this dependency is always present in the Thorntail BOM and so is
always added when adding all artifacts from the BOM.

Result
------
Generating offline Maven repository works fine again.