THORN-2358: add microprofile.jwt aliases to configurables that only use microprofile.jwtauth
Motivation
----------
In THORN-2352, we added 2 new configurables with a prefix
of `thorntail.microprofile.jwtauth`. This prefix is old and
we should use `thorntail.microprofile.jwt`.

However, the old prefix has already made it to a release
(2.4.0.Final), so we need to keep the old names working.

Modifications
-------------
Added aliases for two configuration properties in the MicroProfile
JWT fraction.

Result
------
No behavioral change, just consistency in configuration property
names.